Keep WhatsApp socket seam on existing API
